Add files via upload
About us!

Efficient market hypothesis tells us that an asset price movement is generally the reflection of all the news in the market. Is it true? We don’t know. But, with our chatbot you can know what the market thinks about your wanted stock. 
Problem statement and our solution to it:

Suppose you are a newbie in the field of stock markets or maybe someone too busy to go through all the news in the market about each stock. But it’s your job to know what the market sentiment about certain stocks and with the time constraint it is difficult. Well, if someone has the patience and interest to scrape the news and develop a sentiment analyzer, that’s a different thing. 
We do that for you! No, you don’t have to run the code or fill in huge number of details. Instead, you can reach our chatbot services. Our chatbot, Michael Burry will provide you with the market sentiment of the Nifty 50 stock on simply inputting the name of the stock. Being too young, it is yet to learn how to respond to anything but name of a Nifty 50 stock. 
Is it really useful?

We believe so. For someone who needs quick information or does not have adequate finance or coding background yet can reach Michael Burry to understand the sentiment of a particular stock. 
In the future, algorithmic trading is going to exceed its current reach and we are of the belief that having a sentiment analyzer handy can be useful for adding another constraint as an insurance as well as a brilliant tool for the fund manager to take quick decisions.